Reduce the number of Warnings produced
Make SPL warning free

Restore size_t usage

During a temporary (one hopes) bout of insanity, I had decided that
size_t should be uint32_t and changed all references to it, as well as,
a gringe worthy zsize_t typedef.

This commits reverts all those changes back to upstream `size_t` usage.

One peculiar thing is that the vxproj disables C4242 but it appears to
be "forgotten" halfway through and C4242 Warnings come back. If I use
pragma warning (disable: 4242) individually in each source file it
works. But unsure why the top level project disabling it fails.

+++ ZFSin/spl/include/rpc/xdr.h
@@ -81,7 +81,7 @@ struct xdr_ops {
 
 struct xdr_bytesrec {
 	bool_t xc_is_last_record;
-	uint32_t xc_num_avail;
+	size_t xc_num_avail;
 };
 
 /*
