+++ ZFSin/zfs/module/nvpair/nvpair_alloc_fixed.c
@@ -66,7 +66,7 @@ static int
 nv_fixed_init(nv_alloc_t *nva, va_list valist)
 {
 	uintptr_t base = va_arg(valist, uintptr_t);
-	uintptr_t lim = base + va_arg(valist, uint32_t);
+	uintptr_t lim = base + va_arg(valist, size_t);
 	nvbuf_t *nvb = (nvbuf_t *)P2ROUNDUP(base, sizeof (uintptr_t));
 
 	if (base == 0 || (uintptr_t)&nvb[1] > lim)
@@ -81,7 +81,7 @@ nv_fixed_init(nv_alloc_t *nva, va_list valist)
 }
 
 static void *
-nv_fixed_alloc(nv_alloc_t *nva, uint32_t size)
+nv_fixed_alloc(nv_alloc_t *nva, size_t size)
 {
 	nvbuf_t *nvb = nva->nva_arg;
 	uintptr_t new = nvb->nvb_cur;
@@ -96,7 +96,7 @@ nv_fixed_alloc(nv_alloc_t *nva, uint32_t size)
 
 /*ARGSUSED*/
 static void
-nv_fixed_free(nv_alloc_t *nva, void *buf, uint32_t size)
+nv_fixed_free(nv_alloc_t *nva, void *buf, size_t size)
 {
 	/* don't free memory in the pre-allocated buffer */
 }
